multi_agent_system/requirements.txt
# Core AI and Agent Framework
crewai==0.28.8
langchain>=0.1.10,<0.2.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# OpenAI Integration
openai>=1.13.3,<2.0.0

# Web Scraping and HTTP
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
lxml==4.9.3

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Data Processing and Validation
pydantic==2.5.0
python-dotenv==1.0.0

# Additional Utilities
colorama==0.4.6
rich==13.7.0
typer==0.9.0

# Development Dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
